+++ src/main/java/ru/job4j/service/SimpleVacancyService.java
@@ -1,6 +1,7 @@
 package ru.job4j.service;
 
 import org.springframework.stereotype.Service;
+import ru.job4j.dto.FileDto;
 import ru.job4j.model.Vacancy;
 import ru.job4j.repository.VacancyRepository;
 
@@ -10,26 +11,47 @@ import java.util.Optional;
 @ThreadSafe
 @Service
 public class SimpleVacancyService implements VacancyService {
-
     private final VacancyRepository vacancyRepository;
 
-    public SimpleVacancyService(VacancyRepository vacancyRepository) {
-        this.vacancyRepository = vacancyRepository;
+    private final FileService fileService;
+
+    public SimpleVacancyService(VacancyRepository sql2oVacancyRepository, FileService fileService) {
+        this.vacancyRepository = sql2oVacancyRepository;
+        this.fileService = fileService;
+    }
+    @Override
+    public Vacancy save(Vacancy vacancy, FileDto image) {
+        saveNewFile(vacancy, image);
+        return vacancyRepository.save(vacancy);
     }
 
-    @Override
-    public Vacancy save(Vacancy vacancy) {
-        return vacancyRepository.save(vacancy);
+    private void saveNewFile(Vacancy vacancy, FileDto image) {
+        var file = fileService.save(image);
+        vacancy.setFileId(file.getId());
     }
 
     @Override
     public boolean deleteById(int id) {
-        return vacancyRepository.deleteById(id);
+        var fileOptional = findById(id);
+        if (fileOptional.isEmpty()) {
+            return false;
+        }
+        var isDeleted = vacancyRepository.deleteById(id);
+        fileService.deleteById(fileOptional.get().getFileId());
+        return isDeleted;
     }
 
     @Override
-    public boolean update(Vacancy vacancy) {
-        return vacancyRepository.update(vacancy);
+    public boolean update(Vacancy vacancy, FileDto image) {
+        var isNewFileEmpty = image.getContent().length == 0;
+        if (isNewFileEmpty) {
+            return vacancyRepository.update(vacancy);
+        }
+        var oldFileId = vacancy.getFileId();
+        saveNewFile(vacancy, image);
+        var isUpdated = vacancyRepository.update(vacancy);
+        fileService.deleteById(oldFileId);
+        return isUpdated;
     }
 
     @Override
